refactor searching and filter tx
